Amadeus.Encore: v3.82 network support (#2)
* Amadeus.Encore - Network implementation for NX Client (3.82)

* Update gitignore

* Update resources and music parser

* Identified new payload: User Rank

* Identified new fields in CharacterInfoResponse

* Implemented new endpoint: AcceptGift (`0x139F` / `0x13A0`)

* Move `CreateRoomRequest` to correct directory

* Update README

* Fix log skill detection

* Improved score tracker and room state handling

* Improved codec handling on abstract type

* Amadeus.Encore implementation checkpoint:
- Added support for OJNList.dat
- Added support for Album mode
- Added support for Couple mode
- Proper reward calculation based on OJNList
- Improved metadata resolution

* Update README.md

* Add support returning `Command` instead of `IMessage` as a response

* Add Shop support for v3.82

* Upgrade to .NET 10

Also upgraded dependencies accordingly.
Switched MySQL DB driver to support the migration.

* Add Shop support leftover

* Add missing option

* Release Candidate: 2.0.0rc1

* Add a new extra line at the end of files

* Optimize imports

* Update README.md

* Update README.md

* Update project file

+++ Source/Mozart.Data/Metadata/Items/ItemDataParser.cs
@@ -34,7 +34,7 @@ public static class ItemDataParser
             short bitflag         = reader.ReadInt16();
             item.Gender           = (Gender)((bitflag >> 7) & 15);
             item.IsNew            = (bitflag >> 11) == 1;
-            item.Quantity         = reader.ReadByte(); // reader.ReadInt16();
+            item.Quantity         = reader.ReadInt16();
             item.GameModifier     = (GameModifier) reader.ReadByte();
             item.GameModifierType = (GameModifierType) reader.ReadByte();
             item.Price.Currency   = (Currency)reader.ReadByte();
